refactory and add throws in case of invalid auth
